CRUD methods and exceptions applications

+++ src/main/java/io/github/ngraciano/libraryapi/service/AuthorService.java
@@ -1,8 +1,12 @@
 package io.github.ngraciano.libraryapi.service;
 
 
+import io.github.ngraciano.libraryapi.exceptions.OperationNotPermitted;
 import io.github.ngraciano.libraryapi.model.Author;
 import io.github.ngraciano.libraryapi.repository.AuthorRepository;
+import io.github.ngraciano.libraryapi.repository.BookRepository;
+import io.github.ngraciano.libraryapi.validator.AuthorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,15 +14,15 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AuthorService {
+        private final AuthorValidator validator;
+        private final AuthorRepository repository;
+        private final BookRepository bookRepository;
 
-        private AuthorRepository repository;
-
-    public AuthorService(AuthorRepository repository){
-        this.repository=repository;
-    }
-
-    public Author save(Author author){
+    public Author save(Author author)
+    {
+        validator.Validate(author);
         return repository.save(author);
     }
 
@@ -26,6 +30,7 @@ public class AuthorService {
         if (author.getId()==null){
             throw new IllegalArgumentException("for update,is necessary author saved in the db");
         }
+        validator.Validate(author);
          repository.save(author);
     }
 
@@ -35,6 +40,9 @@ public class AuthorService {
     }
 
     public void delete(Author author){
+        if (hasBook(author)){
+            throw new OperationNotPermitted("Author have book registred");
+        }
          repository.delete(author);
     }
 
@@ -49,5 +57,9 @@ public class AuthorService {
         return repository.findByNationality(nationality);
     }
     return repository.findAll();
+    }
+
+    public boolean hasBook(Author author){
+        return bookRepository.existsByAuthor(author);
     }
 }
